chtml/html/testdata/webkit/math.dat

#data
<math><tr><td><mo><tr>
#errors
#document
| <math math>
|   <math tr>
|     <math td>
|       <math mo>
|         <tr>

#data
<math><tr><td><mo><tr>
#errors
#document
| <math math>
|   <math tr>
|     <math td>
|       <math mo>
|         <tr>

#data
<math><thead><mo><tbody>
#errors
#document
| <math math>
|   <math thead>
|     <math mo>
|       <tbody>

#data
<math><tfoot><mo><tbody>
#errors
#document
| <math math>
|   <math tfoot>
|     <math mo>
|       <tbody>

#data
<math><tbody><mo><tfoot>
#errors
#document
| <math math>
|   <math tbody>
|     <math mo>
|       <tfoot>

#data
<math><tbody><mo></table>
#errors
#document
| <math math>
|   <math tbody>
|     <math mo>

#data
<math><thead><mo></table>
#errors
#document
| <math math>
|   <math thead>
|     <math mo>

#data
<math><tfoot><mo></table>
#errors
#document
| <math math>
|   <math tfoot>
|     <math mo>
